move translator to avoid circular imports
