+++ py/sentry_streams/adapters/stream_adapter.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import Any
+from typing import Any, Optional
 
 from sentry_streams.pipeline import Step
 
@@ -14,3 +14,17 @@ class StreamAdapter(ABC):
     @abstractmethod
     def sink(self, step: Step, stream: Any) -> Any:
         raise NotImplementedError
+
+
+class RuntimeTranslator:
+    def __init__(self, runtime_adapter: StreamAdapter):
+        self.adapter = runtime_adapter
+
+    def translate_step(self, step: Step, stream: Optional[Any] = None) -> Any:
+        assert hasattr(step, "step_type")
+        translated_fn = getattr(self.adapter, step.step_type)
+
+        if stream:
+            return translated_fn(step, stream)
+        else:
+            return translated_fn(step)
